agent/requirements.txt
livekit-agents[openai,turn-detector,silero,gladia]~=1.2
livekit-plugins-noise-cancellation~=0.2
python-dotenv
httpx
pypdf
